refactor: Placing the 'DefaultValueAllowedType' inside field type
